fix(output/http): fix test that timed out
